fix(download_engine): improve torrent metadata handling and resource cleanup
- Add try-finally block to properly manage active torrent IDs and cancel tokens
- Apply file priorities immediately after metadata retrieval
- Recheck existing data on disk to reflect actual progress accurately
- Wait for torrent state to stabilize before resuming and listening for completion
- Close ReceivePort and ErrorPort before spawning new worker in isolate pool
- Cancel response data stream on specific HTTP errors to free resources
- Correct authorization check logic in Remote API service to validate 200 status instead of 401
- Fix integer clamping in torrent progress calculation by adding toInt() conversion

+++ lib/core/services/torrent_service_ffi.dart
@@ -349,7 +349,7 @@ class TorrentService {
             name: f.name,
             size: f.size,
             downloadedBytes: (progress != null && i < progress.length)
-                ? (progress[i] as num).toInt().clamp(0, f.size)
+                ? (progress[i] as num).toInt().clamp(0, f.size).toInt()
                 : 0,
             priority: (priorities != null && i < priorities.length)
                 ? (priorities[i] as num).toInt()
